+++ Source/Alternet.UI.Common/Port/PresentationCore/EventRoute.cs
@@ -17,7 +17,7 @@ namespace Alternet.UI.Port
     /// </summary>
     /// <remarks>
     ///     EventRoute constitues <para/>
-    ///     a non-null <see cref="Alternet.UI.RoutedEvent"/>
+    ///     a non-null <see cref="Alternet.UI.Port.RoutedEvent"/>
     ///     and <para/>
     ///     an ordered list of (target object, handler list)
     ///     pairs <para/>
@@ -27,16 +27,16 @@ namespace Alternet.UI.Port
     ///     and also allows for the handlers in the list 
     ///     to be invoked
     /// </remarks>
-    public sealed class EventRoute
+    internal sealed class EventRoute
     {
         #region Construction
 
         /// <summary>
         ///     Constructor for <see cref="EventRoute"/> given
-        ///     the associated <see cref="Alternet.UI.RoutedEvent"/>
+        ///     the associated <see cref="Alternet.UI.Port.RoutedEvent"/>
         /// </summary>
         /// <param name="routedEvent">
-        ///     Non-null <see cref="Alternet.UI.RoutedEvent"/> to be associated with 
+        ///     Non-null <see cref="Alternet.UI.Port.RoutedEvent"/> to be associated with 
         ///     this <see cref="EventRoute"/>
         /// </param>
         public EventRoute(RoutedEvent routedEvent)